refactoring api routing to match front end on conditions
-renamed condition card file on frontend to be more consistent
-adds medications to returned condition data since the frontend expects it
