+++ client/src/components/ConditionsCard.js
